+++ react/__mocks__/vtex.address-context/index.ts
@@ -1,0 +1,4 @@
+import Utils from './Utils'
+import AddressContext from './AddressContext'
+
+export { Utils, AddressContext }