+++ react/typings/vtex.address-context.d.ts
@@ -6,3 +6,7 @@ declare module 'vtex.address-context/AddressContext' {
 declare module 'vtex.address-context/types' {
   export * from 'vtex.address-context/react/types'
 }
+
+declare module 'vtex.address-context/Utils' {
+  export * from 'vtex.address-context/react/Utils'
+}